+++ packages/cloud-cognitive/src/components/Tearsheet/TearsheetNarrow.stories.js
@@ -11,7 +11,9 @@ import { action } from '@storybook/addon-actions';
 
 import { pkg } from '../../settings';
 import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
-import { getStorybookPrefix } from '../../../config';
+
+import { Button, Form, FormGroup, TextInput } from 'carbon-components-react';
+
 import { Tearsheet, TearsheetNarrow } from '.';
 import {
   actionsOptions,
@@ -19,9 +21,8 @@ import {
   actionsMapping,
 } from '../ActionSet/actions.js';
 
+import { getStorybookPrefix } from '../../../config';
 const storybookPrefix = getStorybookPrefix(pkg, TearsheetNarrow.displayName);
-
-import { Button } from 'carbon-components-react';
 
 import styles from './_storybook-styles.scss';
 
@@ -30,16 +31,11 @@ import mdx from './Tearsheet.mdx';
 export default {
   title: `${storybookPrefix}/Tearsheets/${TearsheetNarrow.displayName}`,
   component: TearsheetNarrow,
-  subcomponents: {
-    Tearsheet,
-  },
+  subcomponents: { Tearsheet },
   parameters: { styles, docs: { page: mdx } },
   argTypes: {
     actions: {
-      control: {
-        type: 'select',
-        labels: actionsLabels,
-      },
+      control: { type: 'select', labels: actionsLabels },
       options: actionsOptions,
       mapping: actionsMapping(
         {
@@ -51,68 +47,34 @@ export default {
         action
       ),
     },
-    className,
     closeIconDescription: {},
-    description: {
-      control: {
-        type: 'text',
-      },
-    },
-    hasCloseIcon: {},
-    height: {},
-    label: {
-      control: {
-        type: 'text',
-      },
-    },
-    preventCloseOnClickOutside: {},
-    title: {
-      control: {
-        type: 'text',
-      },
-    },
-    buttons: {
-      control: {
-        disable: true,
-      },
-    },
-    children: {
-      control: {
-        disable: true,
-      },
-    },
-    onClose: {
-      control: {
-        disable: true,
-      },
-    },
-    open: {
-      control: {
-        disable: true,
-      },
-    },
+    description: { control: { type: 'text' } },
+    label: { control: { type: 'text' } },
+    title: { control: { type: 'text' } },
+    onClose: { control: { disable: true } },
+    open: { control: { disable: true } },
   },
 };
 
 // Test values for props.
 
-const className = 'client-class-1 client-class-2';
-
 const closeIconDescription = 'Close the tearsheet';
 
-const description = (
-  <span>
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-    tempor <strong>incididunt ut labore</strong> et dolore magna aliqua. Ut enim
-    ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip
-    ex ea commodo consequat.
-  </span>
-);
+const description =
+  'This is a description for the tearsheet, providing an opportunity to \
+  describe the flow.';
 
 const label = 'The label of the tearsheet';
 
 const mainContent = (
-  <div className="tearsheet-stories__dummy-content-block">Main content</div>
+  <div className="tearsheet-stories__narrow-content-block">
+    <Form>
+      <p>Main content</p>
+      <FormGroup>
+        <TextInput labelText="Enter an important value here" />
+      </FormGroup>
+    </Form>
+  </div>
 );
 
 const title = 'Title of the tearsheet';
@@ -124,7 +86,7 @@ const Template = ({ actions, ...args }) => {
 
   return (
     <>
-      <style>{`.${pkg.prefix}-tearsheet { opacity: 0 }`};</style>
+      <style>{`.${pkg.prefix}--tearsheet { opacity: 0 }`};</style>
       <Button onClick={() => setOpen(true)}>Open Tearsheet</Button>
       <TearsheetNarrow
         {...args}
@@ -137,17 +99,24 @@ const Template = ({ actions, ...args }) => {
 };
 
 // Stories
-export const AllAttributesSet = Template.bind({});
-AllAttributesSet.args = {
+export const tearsheetNarrow = Template.bind({});
+tearsheetNarrow.storyName = 'Narrow tearsheet';
+tearsheetNarrow.args = {
+  description,
+  onClose: action('onClose called'),
+  title,
+  actions: 6,
+};
+
+export const fullyLoaded = Template.bind({});
+fullyLoaded.storyName = 'Narrow tearsheet with all header items';
+fullyLoaded.args = {
   closeIconDescription,
   description,
   hasCloseIcon: true,
   label,
-  preventCloseOnClickOutside: true,
+  onClose: action('onClose called'),
   title,
   verticalPosition: 'normal',
-  actions: 3,
+  actions: 0,
 };
-
-export const NoAttributesSet = Template.bind({});
-NoAttributesSet.args = {};