+++ packages/cloud-cognitive/src/components/index.js
@@ -19,6 +19,7 @@ export {
   UnauthorizedEmptyState,
 } from './EmptyStates';
 export { ExampleComponent } from './ExampleComponent';
+export { ExportModal } from './ExportModal';
 export { HTTPError403, HTTPError404, HTTPErrorOther } from './HTTPErrors';
 export { ImportModal } from './ImportModal';
 export { ModifiedTabs } from './ModifiedTabs';
@@ -28,7 +29,6 @@ export { RemoveDeleteModal } from './RemoveDeleteModal';
 export { Saving } from './Saving';
 export { SidePanel } from './SidePanel';
 export { StatusIcon } from './StatusIcon';
-export { TagSet } from './TagSet';
 export { Tearsheet, TearsheetNarrow } from './Tearsheet';
 export { UserProfileImage } from './UserProfileImage';
 export { WebTerminal } from './WebTerminal';